webops2243 moved login/logout endpoint to under api/v1
